backend/requirements.txt
Django
djangorestframework==3.8.2
pytz==2018.4
